Pulling apart llvm-sys pieces
